main: switch psplash_main and ui update thread work to fix SDL2 backend
For the SDL2 backend (simulator) the task handler needs to be in the main
thread, otherwise no updates are drawn to the window. Move the lvgl loop
into the main thread/function and the command interface into a thread.

To detect 'QUIT' command, checkup on the command thread in the main loop
and exit if the command thread exited.