gitlab-ci: build splash it for old releases 2.3.x
For OS releases 2.3.x use a older glibc and do not run
the build with 2.4.x SDKs. Provide a build with 2.3.1 SDK
as it has a docker image built. For configuration a newer
meson is needed.

The old SDK does not have the environment set properly, do
it before configuring.