gitlab-ci: force fallback/static linking for libconfig
It is not guaranteed that libconfig is available in the SDK
as static or on the target as dynamic library. Just always
build it and link it statically.

For installation a absolut DESTDIR must be given. In addition
install directly into the 'sysroot' directory, aligning location
with common practice for CI artefacts.